+++ src/leoUI.ts
@@ -579,8 +579,8 @@ export class LeoUI extends NullGui {
         commandBindings.makeAllBindings(this, this._context);
     }
 
-    public showSettings(): void {
-        void this.leoSettingsWebview.openWebview();
+    public showSettings(): Promise<unknown> {
+        return this.leoSettingsWebview.openWebview();
     }
 
     public async put_help(c: Commands, s: string, short_title: string): Promise<void> {
@@ -3981,7 +3981,7 @@ export class LeoUI extends NullGui {
                 }),
                 quickPick.onDidAccept(accepted => {
                     if (/^\d+$/.test(quickPick.value)) {
-                        // * Was an integer
+                        // * Was an integer EASTER EGG
                         this.setupRefresh(Focus.Body,
                             {
                                 tree: true,
@@ -4249,13 +4249,16 @@ export class LeoUI extends NullGui {
                 if (w_truncated) {
                     void vscode.window.showInformationMessage("Truncating headline");
                 }
-
+                if (g.doHook("headkey1", { c: c, p: c.p, ch: '\n', changed: true })) {
+                    return w_p;  // The hook claims to have handled the event.
+                }
                 const undoData = u.beforeChangeHeadline(w_p);
                 c.setHeadString(w_p, p_newHeadline); // Set v.h *after* calling the undoer's before method.
                 if (!c.changed) {
                     c.setChanged();
                 }
                 u.afterChangeHeadline(w_p, 'Edit Headline', undoData);
+                g.doHook("headkey2", { c: c, p: c.p, ch: '\n', changed: true });
                 void this.launchRefresh();
             }
 
@@ -4399,12 +4402,16 @@ export class LeoUI extends NullGui {
         if (!p_name) {
             return value;
         }
+        if (g.doHook("headkey1", { c: c, p: c.p, ch: '\n', changed: true })) {
+            return;  // The hook claims to have handled the event.
+        }
         const undoData = u.beforeChangeHeadline(c.p);
         c.setHeadString(c.p, p_name);  // Set v.h *after* calling the undoer's before method.
         if (!c.changed) {
             c.setChanged();
         }
         u.afterChangeHeadline(c.p, 'Edit Headline', undoData);
+        g.doHook("headkey2", { c: c, p: c.p, ch: '\n', changed: true });
         return value;
     }
 
@@ -6417,8 +6424,7 @@ export class LeoUI extends NullGui {
         title: string,
         message: string,
         yes_all = false,
-        no_all = false,
-
+        no_all = false
     ): Thenable<string> {
         const w_choices = [
             Constants.USER_MESSAGES.YES,