+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 1.0.6
+
+- Fixed goto-global-line command.
+- Implemented c.recursiveImport. 
+- implemented g.execute_shell_commands_with_options.
+- Fixed IdleTime class handler call arguments.
+- Added missing headkey1 and headkey2 plugin hooks.
+- Added Leo's documentation updated for LeoJS with Docusaurus.
+- Fixed debug and breakpoints by setting sourceMapRenames to false in launch.json.
+- Added support for 'FreeMind' file format import.
+
 ## 1.0.5
 
 - Fixed @auto files generation for markdown (.md) files.